requirements.txt
python-telegram-bot==21.3
openai==1.35.5
python-dotenv==1.0.1
schedule==1.2.2
requests==2.31.0